requirements.txt
aiohttp==3.9.5
Flask==3.0.3
numpy==1.26.4
Pillow==9.0.1
Pillow==10.3.0
Requests==2.32.2
